requirements.txt
openai
pinecone-client
black
tiktoken
tqdm
dataclasses_json
streamlit